+++ Tools/rdmp/CommandLine/Gui/ConsoleMainWindow.cs
@@ -8,6 +8,7 @@ using MapsDirectlyToDatabaseTable;
 using MapsDirectlyToDatabaseTable.Revertable;
 using Rdmp.Core.CommandExecution;
 using Rdmp.Core.CommandExecution.AtomicCommands;
+using Rdmp.Core.CommandExecution.AtomicCommands.CatalogueCreationCommands;
 using Rdmp.Core.CommandLine.Gui.Windows;
 using Rdmp.Core.CommandLine.Gui.Windows.RunnerWindows;
 using Rdmp.Core.CommandLine.Options;
@@ -279,6 +280,15 @@ namespace Rdmp.Core.CommandLine.Gui
         private IEnumerable<IAtomicCommand> GetCommands()
         {
 			var o = _treeView.SelectedObject;
+
+			if(ReferenceEquals(o,  Catalogues))
+            {
+				return new IAtomicCommand[] {
+					new ExecuteCommandCreateNewCatalogueByImportingFile(_activator),
+					new ExecuteCommandCreateNewCatalogueByImportingExistingDataTable(_activator),
+				};
+            }
+
             if(o == null)
 				return new IAtomicCommand[0];